Bruh

+++ Practice 4/Models/NationalExamData.cs
@@ -4,13 +4,13 @@ public class NationalExamData
 {
     public uint Id { get; set; }
     public string Year { get; set; } = null!;
-    public uint Mathematics { get; set; }
-    public uint Literature { get; set; }
-    public uint Physics { get; set; }
-    public uint Biology { get; set; }
-    public uint English { get; set; }
-    public uint Chemistry { get; set; }
-    public uint History { get; set; }
-    public uint Geography { get; set; }
-    public uint CivicEducation { get; set; }
+    public float Mathematics { get; set; }
+    public float Literature { get; set; }
+    public float Physics { get; set; }
+    public float Biology { get; set; }
+    public float English { get; set; }
+    public float Chemistry { get; set; }
+    public float History { get; set; }
+    public float Geography { get; set; }
+    public float CivicEducation { get; set; }
 }
